Reorganize project.
