--- src/main/scala/com/dg/advxml/transform/funcs/XmlAction.scala
@@ -1,13 +1,0 @@
-package com.dg.advxml.transform.funcs
-
-import scala.xml.NodeSeq
-
-/**
-  * Adxml
-  * Created by geirolad on 18/06/2019.
-  *
-  * @author geirolad
-  */
-trait XmlAction extends (NodeSeq => NodeSeq){
-	def andThen(that: NodeSeq => NodeSeq) : XmlAction = xml => that(this(xml))
-}
