--- src/main/scala/com/dg/advxml/transform/funcs/XmlFuncsImpls.scala
@@ -1,14 +1,0 @@
-package com.dg.advxml.transform.funcs
-
-import com.dg.advxml.transform.funcs.impls.{Actions, Filters, Zooms}
-
-/**
-  * Adxml
-  * Created by geirolad on 18/06/2019.
-  *
-  * @author geirolad
-  */
-trait XmlFuncsImpls
-  extends Actions
-    with Zooms
-    with Filters
